Fix build for newer GCC/arm-eabi toolchain
- board: undef CONFIG_LTO (memcpy link error with GCC 14)
- rom_chip.h: cast 0x40 to uintptr_t for array-bounds warning
- system.c: suppress -Warray-bounds for ROM API
- ecst.c: fix NULL check for hdr_args (GCC strictness)
- ectool.c: initialize ec_response_sysinfo struct
- iteflash.c: suppress deprecated ftdi_usb_purge_buffers, fix strsignal

+++ chip/npcx/rom_chip.h
@@ -42,7 +42,7 @@ enum API_RETURN_STATUS_T {
 /*
  * Macro functions of ROM api functions
  */
-#define ADDR_DOWNLOAD_FROM_FLASH (*(volatile uint32_t *) 0x40)
+#define ADDR_DOWNLOAD_FROM_FLASH (*(volatile uint32_t *)(uintptr_t)0x40)
 #define download_from_flash(src_offset, dest_addr, size, sign, exe_addr, \
 	status) \
 	(((download_from_flash_ptr) ADDR_DOWNLOAD_FROM_FLASH) \
